requirements.txt
beautifulsoup4
google-cloud-texttospeech
lxml
pandas
requests
tqdm
